src/server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12, <3.13"
dependencies = [
    "fastapi",
    "opentelemetry-api",
    "opentelemetry-sdk",
    "opentelemetry-instrumentation-fastapi",
    "opentelemetry-exporter-gcp-trace",
    "pydantic",
    "redis>=5.0.0",
    "uvicorn",
    "httpx",
    "tqdm>=4.50",
]

[project.optional-dependencies]
cpu = [
    "datasets",
    "numpy",
    "peft",
    "torch",
    "transformers>=5.5.4",
]
gpu = [
    "datasets",
    "numpy",
    "peft",
    "torch; sys_platform == 'linux'",
    "transformers>=5.5.4",
]
vllm = [
    "torch; sys_platform == 'linux'",
    "torch-c-dlpack-ext; sys_platform == 'linux'",
    "vllm==0.19.1; sys_platform == 'linux'",
]

[tool.uv]
override-dependencies = [
    # vLLM still carries an older transformers constraint; keep the project on
    # the newer transformers line declared in our extras.
    "transformers>=5.5.4",
]
conflicts = [
    [
        { extra = "cpu" },
        { extra = "gpu" },
    ],
    [
        { extra = "cpu" },
        { extra = "vllm" },
    ],
]

[[tool.uv.index]]
name = "torch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "torch-gpu"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
# Pick the torch build from the user-facing cpu/gpu extras; vllm always uses CUDA torch.
torch = [
    { index = "torch-cpu", extra = "cpu" },
    { index = "torch-gpu", extra = "gpu", marker = "sys_platform == 'linux'" },
    { index = "torch-gpu", extra = "vllm", marker = "sys_platform == 'linux'" },
]
